Edit booking (#13)
* Create CustomerCompletedBooking.vue

Added new customer completed booking history component. This is to separate completed bookings from pending ones.

* Update CustomerHistory.vue

Updated the component to accommodate the changes to separate completed and pending bookings.

* Create UpdateBooking.vue

Added new UpdateBooking component to allow the customer to update their booking(s) if the departure date of the booking is 3 days or more.

* Update BookingForm.vue

Added a confirmation modal before the customer creates a booking.

* Update CustomerBookingHistory.vue

Updated the view to accommodate the changes of separating pending and completed bookings.